+++ src/thrift_gen/mod.rs
@@ -1,1 +1,2 @@
 pub mod binary_format;
+pub mod zipkin_core;